Made some fairly significant code (but not functionality) changes in Track, which percolated out pretty far.
Previously, Track had held only indices into a detection vector; it was up to the user to extract DiaIds later.  The user actually had a public member variable componentDetectionIds which they 
could modify as they pleased.  

This has been changed so that the Right Way to add detections to a track is addDetection(detIndex, detectionVector).  It will get both the index into the vector (used by linkTracklets) as well 
as the diaIds themselves (used in output, and probably more interesting to humans.)  I changed around linkTracklets as well as the unit tests to reflect this and got the unit tests 
running.  

Next I plan to remove the public member variables entirely and require interactions via addDetection() and getters only.  Perhaps I will add a removeDetection/setters as well if we find need for 
it.  I will actually merge these changes back to trunk when I'm done because there is no reason not to do so.