Updated distributedLinkTracklets code to use new functions which are (almost) identical to those from the branched version of the trunk.
The new version does NOT, however, calculate per-track RMS but does track filtering the old fashioned way: it first checks the quality of fit of the 
endpoint tracklets' detections to the best-fit function, then adds support points if they are sufficiently close. This is to maintain consistency 
between the assumptions made in Matt's original computational model and the code in SVN.  After his thesis is over, we will reconsider the 
computational model in use and evaluate the per-track RMS fitting.

The code is now compiling, but has NOT been tested.