Ticket 1578 is already closed, but I tried an experiment here and it paid off big time!
It appears the secret to boosting our performance was in tree creation; previously when partitioning data, we chose our pivot, then gave the left 
child the bounds (min, pivot) and the right child (pivot, max).  Now, we actually recalculate the min/max at the children.  The subtle difference is 
that now if there is no data betweent he child's max element and pivot, the tree does not keep that region of space in its bounds.  This makes a big 
difference at the bottom level of the tree - nodes are actually much smaller, especially at the bottom level.

Our performance is now through the roof and #tracks/#found objects close to the C version.  Interestingly, though, we fail some of my unit tests.

A handful of other changes are committed here as well since my discovery came while attempting a larger overhaul.  Among them:

 - We now rotate all data to center on (180,0)
 - TrackletTree no longer inherits a constructor but does its own construction.
 - TrackletTree/TrackletTreeNode takes a "widths" array at construction time like the C TBT_ARRAY.
 - Currently "widths" are calculated by default in TrackletTree and set to the width of the given image's tracklets in each dimension (maxRA - minRa, 
etc) - note that in C, the width of the *full set of tracklets* is used, after the addition of error in each dimension.
 - TrackletTreeNode by default will choose the average of the min/max as the pivot rather than the median element when splitting data in a given 
dimension. Old behavior can be chosen optionally.
 - TrackletTreeNode by default will not split each dimension sequentially (RA, Dec, RaV, DecV) but at each iteration chooses the "widest" dimension 
to split.